[HOTSPOT-188] refactor: 구조 및 import 경로 수정
