+++ apps/web/src/entities/policy-order/index.ts
@@ -1,2 +1,2 @@
-export type { PolicyOrderType } from './model/type';
+export type { FamilyPriority, PolicyOrderType } from './model/type';
 export { POLICY_ORDER_TYPE } from './model/type';
