+++ apps/web/src/features/policy-apply/index.ts
@@ -1,2 +1,3 @@
-export { BlockAddList } from './ui/BlockAddList';
+export type { PolicyApply } from './model/types';
+export { useApplyPolicy } from './model/useApplyPolicy';
 export { PolicyAddList } from './ui/PolicyAddList';
