+++ apps/web/src/features/policy-order/index.ts
@@ -1,2 +1,3 @@
+export { useFifoOrder } from './model/useFifoOrder';
 export { PolicyOrderSelector } from './ui/PolicyOrderSelector';
 export { PolicyPriorityList } from './ui/PolicyPriorityList';
